+++ src/box.tsx
@@ -1,16 +1,9 @@
 import * as React from 'react';
-import {createUseStyles} from './jss';
 import classnames from 'classnames';
+import {sprinkles} from './sprinkles.css';
 import {getPrefixedDataAttributes} from './utils/dom';
 
 import type {DataAttributes} from './utils/types';
-
-const useStyles = createUseStyles(() => ({
-    box: {
-        padding: (p) => p.padding,
-        width: (p) => p.width,
-    },
-}));
 
 type PadSize = 0 | 2 | 4 | 8 | 12 | 16 | 24 | 32 | 40 | 48 | 56 | 64 | 72 | 80;
 
@@ -50,18 +43,20 @@ const Box = React.forwardRef<HTMLDivElement, Props>(
         },
         ref
     ) => {
-        const classes = useStyles({
-            padding: `${paddingTop}px ${paddingRight}px ${paddingBottom}px ${paddingLeft}px`,
-            width,
-        });
-
         return (
             <div
                 {...getPrefixedDataAttributes(dataAttributes)}
-                className={classnames(className, classes.box)}
                 role={role}
                 aria-label={ariaLabel}
                 ref={ref}
+                className={classnames(
+                    className,
+                    sprinkles({paddingTop, paddingBottom, paddingLeft, paddingRight})
+                )}
+                style={{
+                    width,
+                    boxSizing: 'border-box',
+                }}
             >
                 {children}
             </div>