+++ src/test-utils/fail-test-on-console-error.tsx
@@ -10,7 +10,10 @@ beforeEach(() => {
 
     jest.spyOn(global.console, 'warn').mockImplementation((...args) => {
         // There is a known issue in JSS that makes this warning to trigger when rendering in Node
-        if (args[0].includes('[JSS] Rule is not linked. Missing sheet option "link: true"')) {
+        if (
+            args[0].includes('[JSS] Rule is not linked. Missing sheet option "link: true"') ||
+            args[0].includes('[JSS] CSSOM.parse is not a function')
+        ) {
             return;
         }
         errorOrWarnCalled = true;