+++ src/theme.tsx
@@ -6,7 +6,7 @@ import type {Skin, Colors, SkinName, TextPresetsConfig} from './skins/types';
 import type {TrackingEvent} from './utils/types';
 import type {MediaQueries} from './utils/media-queries';
 
-export type ThemeTexts = typeof TEXTS_ES;
+export type ThemeTexts = Readonly<typeof TEXTS_ES>;
 
 const TEXTS_ES = {
     expirationDatePlaceholder: 'MM/AA',
@@ -160,17 +160,15 @@ export const getTexts = (locale: Locale): typeof TEXTS_ES => {
     }
 };
 
+export const NAVBAR_HEIGHT_MOBILE = 56;
+export const NAVBAR_HEIGHT_DESKTOP = 80;
+
 export const dimensions = {
-    headerMobileHeight: 56,
-    headerDesktopHeight: 80,
+    headerMobileHeight: NAVBAR_HEIGHT_MOBILE,
+    headerDesktopHeight: NAVBAR_HEIGHT_DESKTOP,
 };
 
-export const mediaQueriesConfig = {
-    tabletMinWidth: 768,
-    desktopMinWidth: 1024,
-    largeDesktopMinWidth: 1368,
-    desktopOrTabletMinHeight: 550,
-};
+export {mediaQueriesConfig} from './media-queries.css';
 
 type LinkComponent = React.ComponentType<{
     style?: React.CSSProperties;
@@ -212,34 +210,38 @@ export type EventFormat = 'universal-analytics' | 'google-analytics-4';
 
 // This is the type expected by ThemeContextProvider theme prop.
 // This config is provided by the user of the lib
-export type ThemeConfig = {
+export type ThemeConfig = Readonly<{
     skin: Readonly<Skin>;
     colorScheme?: ColorScheme; // light by default. TODO: Change to auto by default in next major version
-    i18n: {
+    i18n: Readonly<{
         locale: Locale;
         phoneNumberFormattingRegionCode: RegionCode;
-    };
-    platformOverrides?: {
+    }>;
+    platformOverrides?: Readonly<{
         platform?: 'ios' | 'android' | 'desktop';
         insideNovumNativeApp?: boolean;
         userAgent?: string;
-    };
+    }>;
     texts?: Partial<ThemeTexts>;
-    analytics?: {
+    analytics?: Readonly<{
         logEvent: (trackingEvent: TrackingEvent) => Promise<void>;
         eventFormat?: EventFormat;
-    };
-    dimensions?: {headerMobileHeight: number};
-    mediaQueries?: {
+    }>;
+    dimensions?: Readonly<{headerMobileHeight: number | 'mistica'}>;
+
+    /**
+     * @deprecated in future versions, mistica won't allow to configure media query breakpoints.
+     */
+    mediaQueries?: Readonly<{
         tabletMinWidth: number;
         desktopMinWidth: number;
         largeDesktopMinWidth: number;
         desktopOrTabletMinHeight: number;
-    };
+    }>;
     Link?: LinkComponent;
     useHrefDecorator?: () => (href: string) => string;
     enableTabFocus?: boolean;
-};
+}>;
 
 // This is the lib INTERNAL context
 export type Theme = {