+++ src/touchable.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import classnames from 'classnames';
 import ScreenReaderOnly from './screen-reader-only';
-import {createUseStyles} from './jss';
+import * as classes from './touchable.css';
 import {useTheme} from './hooks';
 import {isInsideNovumNativeApp} from './utils/platform';
 import {ENTER, SPACE} from './utils/key-codes';
@@ -19,41 +19,6 @@ const redirect = (url: string, external = false, loadOnTop = false): void => {
         document.location.href = url;
     }
 };
-
-const useStyles = createUseStyles(() => ({
-    touchable: {
-        color: 'inherit',
-        verticalAlign: 'bottom', // required to remove bottom gap when rendered as inline-block div
-        fontFamily: 'inherit',
-        overflow: 'visible',
-        appearance: 'none',
-        border: 'none',
-        cursor: 'pointer',
-        display: 'block',
-        userSelect: 'none',
-        backgroundColor: 'transparent',
-        padding: 0,
-        textAlign: 'inherit',
-        textDecoration: 'none',
-        fontSize: 'inherit',
-        WebkitTapHighlightColor: 'transparent',
-        width: '100%',
-        '&::-moz-focus-inner': {
-            padding: 0,
-            border: 'none',
-        },
-        '&[disabled]': {
-            cursor: 'default',
-        },
-        '&:active, &:hover': {
-            textDecoration: 'none',
-        },
-    },
-
-    notTouchable: {
-        cursor: 'auto',
-    },
-}));
 
 export type PressHandler = (event: React.MouseEvent<HTMLElement>) => void;
 
@@ -131,10 +96,9 @@ export interface PropsMaybeOnPress extends CommonProps {
 export type Props = PropsHref | PropsTo | PropsOnPress | PropsMaybeHref | PropsMaybeTo | PropsMaybeOnPress;
 export type TouchableElement = HTMLDivElement | HTMLAnchorElement | HTMLButtonElement;
 
-const Touchable = React.forwardRef<TouchableElement, Props>((props, ref) => {
+const RawTouchable = React.forwardRef<TouchableElement, Props>((props, ref) => {
     const {texts, analytics, platformOverrides, Link, useHrefDecorator} = useTheme();
     const hrefDecorator = useHrefDecorator();
-    const classes = useStyles();
     const isClicked = React.useRef(false);
     let trackingEvents: ReadonlyArray<TrackingEvent> = [];
     if (props.trackingEvent) {
@@ -148,7 +112,7 @@ const Touchable = React.forwardRef<TouchableElement, Props>((props, ref) => {
     const children = props.children;
 
     const commonProps = {
-        className: classnames(classes.touchable, props.className),
+        className: props.className,
         disabled: props.disabled,
         style: props.style,
         role: props.role,
@@ -297,6 +261,15 @@ const Touchable = React.forwardRef<TouchableElement, Props>((props, ref) => {
             {children}
         </div>
     );
+});
+
+const Touchable = React.forwardRef<TouchableElement, Props>((props, ref) => {
+    return <RawTouchable {...props} className={classnames(classes.touchable, props.className)} ref={ref} />;
+});
+
+// Used internally by Mistica's components to avoid styles collisions
+export const BaseTouchable = React.forwardRef<TouchableElement, Props>((props, ref) => {
+    return <RawTouchable {...props} className={classnames(classes.base, props.className)} ref={ref} />;
 });
 
 export default Touchable;